saaj_backend_python/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
boto3==1.35.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
aiosmtplib==3.0.2
razorpay==1.4.2
PyJWT==2.9.0
cryptography==43.0.1
slowapi==0.1.9
